requirements.txt
asgiref==3.9.1
cffi==2.0.0
cryptography==45.0.7
Django==5.2.6
djangorestframework==3.15.2
djangorestframework_simplejwt[blacklist]==5.5.1
drf-yasg==1.21.10
gunicorn==21.2.0
psycopg2-binary==2.9.10
pycparser==2.23
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2